+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [0.0.2] - 2025-11-22
+
+### Added
+- User login/logout functionality
+- PostgreSQL DB container
 
 ## [0.0.1] - 2025-11-22